+++ kite-core/src/main/kotlin/com/tang/kite/sql/Column.kt
@@ -3,8 +3,9 @@ package com.tang.kite.sql
 import com.tang.kite.constants.SqlString
 import com.tang.kite.function.SFunction
 import com.tang.kite.utils.Reflects
+import java.io.Serializable
 import java.lang.reflect.Field
-import kotlin.reflect.KMutableProperty1
+import kotlin.reflect.KProperty1
 import kotlin.reflect.jvm.javaField
 
 /**
@@ -18,11 +19,11 @@ open class Column(
 
     var tableAlias: String? = null
 
-) {
+) : Serializable {
 
     constructor(field: Field) : this(Reflects.getColumnName(field), field)
 
-    constructor(column: KMutableProperty1<*, *>): this(column.javaField!!)
+    constructor(column: KProperty1<*, *>): this(column.javaField!!)
 
     constructor(column: SFunction<*, *>): this(Reflects.getField(column))