+++ kite-core/src/main/kotlin/com/tang/kite/sql/function/SqlKeyword.kt
@@ -1,0 +1,30 @@
+package com.tang.kite.sql.function
+
+/**
+ * Pure SQL syntax keywords
+ *
+ * @author Tang
+ */
+object SqlKeyword {
+
+    const val STAR = "*"
+    const val FROM = "from"
+    const val TRIM_BOTH = "both"
+    const val TRIM_LEADING = "leading"
+    const val TRIM_TRAILING = "trailing"
+    const val SEPARATOR = "separator"
+    const val EXISTS = "exists"
+    const val IN = "in"
+    const val DISTINCT = "distinct"
+    const val ALL = "all"
+
+    const val DATE = "date"
+    const val TIME = "time"
+    const val YEAR = "year"
+    const val MONTH = "month"
+    const val DAY = "day"
+    const val HOUR = "hour"
+    const val MINUTE = "minute"
+    const val SECOND = "second"
+
+}